* Deprecate gd_bit_t and gd_spf_t; delete gd_count_t.  The former two are still defined, but no longer used in the API. * the count_max parameter to MPLEX is optional. * Deprecate GD_FLOAT and GD_DOUBLE, and the corresponding symbols in the Standards. * Finally finished _GD_FiniRawIO. * Allow the caller to modify the MPLEX lookback length via gd_mplex_lookback(). * gd_tokenise -> gd_strtok to sidestep regional spelling varations. * Added new gd_entry_list() and gd_nentries() and then made all the other counting and list functions special cases of these. * Update mans.

+++ test/alias_list.c
@@ -56,7 +56,7 @@ int main(void)
   if (!r)
     for (i = 0; al[i]; ++i) {
       CHECKUi(i, al[i][1], 0);
-      x[al[i][0]]++;
+      x[(int)(al[i][0])]++;
     }
 
   CHECKI(i, 6);
